refactor(events): split security debate runner into aragora.debate (P4a Batch E7a)
Relocate trigger_security_debate, _get_security_debate_agents, and
build_security_debate_question from aragora.events.security_events to a
new aragora/debate/security_response.py. These functions import Arena,
DebateProtocol, and API agents, so they are the domain-coupled half of
the security events module.

- Add register_security_debate_runner/get_security_debate_runner to
  aragora.events.security_events; invert SecurityEventEmitter's
  _trigger_security_debate to call the registered runner instead of
  importing the debate function directly.
- Invert SecurityDispatcher's default _run_debate path to route through
  the same registered runner instead of importing
  aragora.debate.orchestrator.Arena directly.
- Drop the dead 'from aragora.agents.factory import get_available_agents'
  fallback in the relocated _get_security_debate_agents (that module does
  not exist).
- aragora/debate/orchestrator.py imports aragora.debate.security_response
  for its self-registration side effect, guaranteeing the runner is
  registered before any Arena-backed process can emit a security event.
- Repoint aragora/debate/security_debate.py and all test @patch targets
  (tests/debate/test_security_debate.py, new
  tests/debate/test_security_response.py) to the new module; no
  re-export shim per the Relocate-Up exemption.
- Add default-path regression tests to tests/events/test_security_dispatcher.py
  proving the dispatcher no longer needs to import Arena directly.
- Hand-shrink scripts/baselines/import_contracts_baseline.json: remove
  the now-cleared 'aragora.events -> aragora.agents' edge only (E7a is
  its sole contributor); 'aragora.events -> aragora.debate' is kept
  since E7b has not landed yet.
