+++ aragora/debate/security_debate.py
@@ -58,7 +58,7 @@ async def run_security_debate(
     from aragora.core_types import DebateResult, Environment
     from aragora.debate.orchestrator import Arena
     from aragora.debate.protocol import DebateProtocol
-    from aragora.events.security_events import build_security_debate_question
+    from aragora.debate.security_response import build_security_debate_question
 
     # Build the debate question from security findings
     question = build_security_debate_question(event)
