+++ scripts/baselines/import_contracts_baseline.json
@@ -37,9 +37,10 @@
         "aragora.debate -> aragora.workflow",
         "aragora.errors -> aragora.server",
         "aragora.evaluation -> aragora.server",
-        "aragora.events -> aragora.agents",
         "aragora.events -> aragora.debate",
+        "aragora.events -> aragora.nomic",
         "aragora.events -> aragora.server",
+        "aragora.events -> aragora.workflow",
         "aragora.evidence -> aragora.connectors",
         "aragora.exceptions -> aragora.connectors",
         "aragora.exceptions -> aragora.server",
